+++ Sources/StreamChatSwiftUI/ChatMessageList/MessageAttachmentsView.swift
@@ -165,12 +165,12 @@ enum MessageAttachmentsBubbleConfiguration {
 
 private extension ChatMessage {
     var hasSingleFileOrVoiceAttachmentWithoutCaption: Bool {
-        guard text.isEmpty else { return false }
+        guard text.isEmpty, quotedMessage == nil else { return false }
         return attachmentCounts.count == 1 && (attachmentCounts[.file] == 1 || attachmentCounts[.voiceRecording] == 1)
     }
     
     var hasSingleMediaAttachmentWithoutCaption: Bool {
-        guard text.isEmpty else { return false }
+        guard text.isEmpty, quotedMessage == nil else { return false }
         return attachmentCounts.count == 1 && (attachmentCounts[.image] == 1 || attachmentCounts[.video] == 1)
     }
 }